+++ packages/cs-shared/formatByPattern.js
@@ -40,7 +40,7 @@ const numberPatterns = {
 /**
  * lang to ICU locale hook (if defined by setLang2LocaleHook)
  * @private
- * @type {null|function}
+ * @type {null|Function}
  */
 let l2lHook = null
 const langToICU = {
@@ -60,8 +60,9 @@ if (typeof _defaultLang === 'undefined') {
 
 /**
  * Create a ICU locale based on UB language
+ *
  * @param lang
- * @return {string}
+ * @returns {string}
  */
 function lang2locale (lang) {
   if (l2lHook) return l2lHook(lang)
@@ -78,6 +79,7 @@ function lang2locale (lang) {
  *
  * Keys is a language, values is an object with keys is date pattern, value is Intl.NumberFormat for this pattern
  * {en: {sum: new Intl.NumberFormat('en-US', numberPatterns.sum)}
+ *
  * @private
  */
 let numberFormaters = {}
@@ -87,27 +89,28 @@ let numberFormaters = {}
  *
  * Keys is a language, values is an object with keys is date pattern, value is Intl.DateTimeFormat for this pattern
  * {en: {date: new Intl.DateTimeFormat('en-US', datePatterns.date)}
+ *
  * @private
  */
 let dateTimeFormaters = {}
 
 /**
  * Format date by pattern
- * @example
-    const formatByPattern = require('@unitybase/cs-shared').formatByPattern
-    const d = new Date(2020, 04, 23, 13, 14)
-    formatByPattern.formatDate(d, 'date') // on client can be called without 3rd lang parameter - will be formatted for user default lang (for uk - 23.05.2020)
-    formatByPattern.formatDate('2020-05-23', 'date', 'uk') // 23.05.2020
-    formatByPattern.formatDate(d, 'date', 'en') // 05/23/2020
-    formatByPattern.formatDate(d, 'dateTime', 'uk') // 23.05.2020 13:14
-    formatByPattern.formatDate(d, 'dateTimeFull', 'uk') // 23.05.2020 13:14:00
-    formatByPattern.formatDate(d, 'date', 'en') // 05/23/2020, 1:14 PM
  *
+ * @example
+const formatByPattern = require('@unitybase/cs-shared').formatByPattern
+const d = new Date(2020, 04, 23, 13, 14)
+formatByPattern.formatDate(d, 'date') // on client can be called without 3rd lang parameter - will be formatted for user default lang (for uk - 23.05.2020)
+formatByPattern.formatDate('2020-05-23', 'date', 'uk') // 23.05.2020
+formatByPattern.formatDate(d, 'date', 'en') // 05/23/2020
+formatByPattern.formatDate(d, 'dateTime', 'uk') // 23.05.2020 13:14
+formatByPattern.formatDate(d, 'dateTimeFull', 'uk') // 23.05.2020 13:14:00
+formatByPattern.formatDate(d, 'date', 'en') // 05/23/2020, 1:14 PM
  * @param {*} dateVal Date object or Number/String what will be converted to Date using new Date();
  *   null, undefined and empty string will be converted to empty string
  * @param {string} patternName One of `formatByPattern.datePatterns`
  * @param {string} [lang=defaultLang] UB language code. If not specified value defined by setDefaultLang is used
- * @return {string}
+ * @returns {string}
  */
 module.exports.formatDate = function (dateVal, patternName, lang = _defaultLang) {
   if (!dateVal) return ''
@@ -126,17 +129,17 @@ module.exports.formatDate = function (dateVal, patternName, lang = _defaultLang)
 
 /**
  * Format number by pattern. Use parseFloat to convert non-number numVal argument into Number. Returns empty string for `!numVal` and `NaN`
- * @example
- const formatByPattern = require('@unitybase/cs-shared').formatByPattern
- const n = 2305.1
- formatByPattern.formatNumber(n, 'sum', 'en') // 2,305.10
- formatByPattern.formatNumber('2305.1', 'sum', 'en') // 2,305.10
- formatByPattern.formatNumber(n, 'sum') // on client can be called without 3rd lang parameter - will be formatted for user default lang (for uk "2 305,10")
  *
+ * @example
+const formatByPattern = require('@unitybase/cs-shared').formatByPattern
+const n = 2305.1
+formatByPattern.formatNumber(n, 'sum', 'en') // 2,305.10
+formatByPattern.formatNumber('2305.1', 'sum', 'en') // 2,305.10
+formatByPattern.formatNumber(n, 'sum') // on client can be called without 3rd lang parameter - will be formatted for user default lang (for uk "2 305,10")
  * @param {*} numVal
  * @param {string} patternName One of `formatByPattern.datePatterns`
  * @param {string} [lang=defaultLang] UB language code. If not specified value defined by `setDefaultLang` is used
- * @return {string}
+ * @returns {string}
  */
 module.exports.formatNumber = function (numVal, patternName, lang = _defaultLang) {
   if (!numVal && (numVal !== 0)) return ''
@@ -159,7 +162,7 @@ module.exports.formatNumber = function (numVal, patternName, lang = _defaultLang
  *
  * Application can redefine this rule by sets his own hook, for example to translate `en -> 'en-GB'` etc.
  *
- * @param {function} newL2lHook function whats takes a UB language string and returns a ICU locale string
+ * @param {Function} newL2lHook function what takes a UB language string and returns a ICU locale string
  */
 module.exports.setLang2LocaleHook = function (newL2lHook) {
   l2lHook = newL2lHook
@@ -169,25 +172,26 @@ module.exports.setLang2LocaleHook = function (newL2lHook) {
 }
 /**
  * Available date patterns
+ *
  * @type {string[]}
  */
 module.exports.datePatterns = Object.keys(datePatterns)
 /**
  * Available Number patterns
+ *
  * @type {string[]}
  */
 module.exports.numberPatterns = Object.keys(numberPatterns)
 
 /**
  * Registers custom date pattern (should be called once)
+ *
  * @example
-
- // format Date for New_York time zone
- $UB.formatter.registerDatePattern('dateTimeInNewYork', {
-    month: '2-digit', day: 'numeric', year: 'numeric', hour: '2-digit', minute: '2-digit',
-    timeZone: 'America/New_York'
- })
-
+// format Date for New_York time zone
+$UB.formatter.registerDatePattern('dateTimeInNewYork', {
+  month: '2-digit', day: 'numeric', year: 'numeric', hour: '2-digit', minute: '2-digit',
+  timeZone: 'America/New_York'
+})
  * @param {string} patternName Pattern name
  * @param {object} intlOptions Intl.DateFormat constructor options - see https://tc39.es/ecma402/#datetimeformat-objects
  */
@@ -206,6 +210,7 @@ module.exports.registerDatePattern = function (patternName, intlOptions) {
 
 /**
  * Registers custom number pattern
+ *
  * @param {string} patternName Pattern name
  * @param {object} intlOptions Intl.NumberFormat constructor options - see https://tc39.es/ecma402/#numberformat-objects
  */
@@ -224,6 +229,7 @@ module.exports.registerNumberPattern = function (patternName, intlOptions) {
 
 /**
  * Gets date pattern by name
+ *
  * @param {string} patternName Pattern name
  * @returns {object} Pattern description for Intl
  */
@@ -233,6 +239,7 @@ module.exports.getDatePattern = function (patternName) {
 
 /**
  * Gets number pattern by name
+ *
  * @param {string} patternName Pattern name
  * @returns {object} Pattern description for Intl
  */
@@ -242,7 +249,8 @@ module.exports.getNumberPattern = function (patternName) {
 
 /**
  * Set a default language to use with `strCmp`, `formatNumber` and `formatDate`.
- * For UI this is usually a logged in user language.
+ * For UI this is usually a logged in user language
+ *
  * @param {string} lang
  */
 function setDefaultLang (lang) {
@@ -260,16 +268,34 @@ function setDefaultLang (lang) {
 module.exports.setDefaultLang = setDefaultLang
 
 /**
- * Compare two value using `Intl.collator` for default language.
- * Returns 0 if values are equal, otherwise 1 or -1.
+ * Compare two value:
+ *  - if one of value is string takes into account current client locale (see setDefaultLang)
+ *  - `null` and `undefined` is always less when any other value
+ *  - `Date` objects compared correctly (using getTime())
+ * Returns 0 if values are equal, otherwise 1 or -1
+ *
  * @param {*} v1
  * @param {*} v2
- * @return {number}
+ * @returns {number}
  */
 module.exports.collationCompare = function (v1, v2) {
-  if (_collator) {
+  if (_collator && ((typeof v1 === 'string') || (typeof v2 === 'string'))) { // Use collator for strings
     return _collator.compare(v1, v2)
   } else {
+    // place null and undefined first (== is used instead of === to null\undefined compare)
+    if ((v1 == null) && (v2 != null)) {
+      return -1
+    } else if ((v2 == null) && (v1 != null)) {
+      return 1
+    }
+    // compare date using seconds since epoch
+    if (v1 instanceof Date) {
+      v1 = v1.getTime()
+    }
+    if (v2 instanceof Date) {
+      v2 = v2.getTime()
+    }
+
     if (v1 === v2) return 0
     return v1 > v2 ? 1 : -1
   }